[master] Laizy loading, preload strategy, is auth build errors
The header and home are moved into a core module, and the recipes are loaded via laizy loading. The auth direct isAuthenticated method is changed into the pages and now is used only a method which calls the auth.isAuthenticated method. Added a new strategy of loading, called preloadingStrategy: PreloadAllModules